Support writing stream to stdout with -O - as wget
Example:

  gdown https://drive.google.com/uc?id=0B9P1L--7Wd2vQm9FaUFHa3h4bFE -O - -q | tar zxvf -

+++ gdown/__init__.py
@@ -12,6 +12,7 @@ import sys
 import tempfile
 
 import requests
+import six
 import tqdm
 
 
@@ -79,37 +80,48 @@ def download(url, output, quiet):
         else:
             output = osp.basename(url)
 
-    if not quiet:
-        print('Downloading...')
-        print('From: %s' % url_origin)
-        print('To: %s' % osp.abspath(output))
+    output_is_path = isinstance(output, six.string_types)
 
-    tmp_file = tempfile.mktemp(
-        suffix=tempfile.template,
-        prefix=osp.basename(output),
-        dir=osp.dirname(output),
-    )
+    if not quiet:
+        print('Downloading...', file=sys.stderr)
+        print('From:', url_origin, file=sys.stderr)
+        print('To:', osp.abspath(output) if output_is_path else output,
+              file=sys.stderr)
+
+    if output_is_path:
+        tmp_file = tempfile.mktemp(
+            suffix=tempfile.template,
+            prefix=osp.basename(output),
+            dir=osp.dirname(output),
+        )
+        f = open(tmp_file, 'wb')
+    else:
+        tmp_file = None
+        f = output
+
     try:
-        with open(tmp_file, 'wb') as f:
-            total = res.headers.get('Content-Length')
-            if total is not None:
-                total = int(total)
+        total = res.headers.get('Content-Length')
+        if total is not None:
+            total = int(total)
+        if not quiet:
+            pbar = tqdm.tqdm(total=total, unit='B', unit_scale=True)
+        chunk_size = 1024  # bytes
+        for chunk in res.iter_content(chunk_size=chunk_size):
+            f.write(chunk)
             if not quiet:
-                pbar = tqdm.tqdm(total=total, unit='B', unit_scale=True)
-            chunk_size = 1024  # bytes
-            for chunk in res.iter_content(chunk_size=chunk_size):
-                f.write(chunk)
-                if not quiet:
-                    pbar.update(len(chunk))
-            if not quiet:
-                pbar.close()
-        shutil.copy(tmp_file, output)
+                pbar.update(len(chunk))
+        if not quiet:
+            pbar.close()
+        if tmp_file:
+            f.close()
+            shutil.copy(tmp_file, output)
     except IOError as e:
         print(e, file=sys.stderr)
         return
     finally:
         try:
-            os.remove(tmp_file)
+            if tmp_file:
+                os.remove(tmp_file)
         except OSError:
             pass
 
@@ -141,6 +153,9 @@ def main():
     parser.add_argument('--id', action='store_true',
                         help='flag to specify file id instead of url')
     args = parser.parse_args()
+
+    if args.output == '-':
+        args.output = sys.stdout.buffer
 
     if args.id:
         url = 'https://drive.google.com/uc?id={id}'.format(id=args.url_or_id)
